edit Components: InputField, PrimaryButton, SocialButton and Title
