+++ mobile/components/InputField.tsx
@@ -1,0 +1,67 @@
+import {StyleSheet, Text, TextInput, View} from "react-native";
+import {Ionicons} from "@expo/vector-icons";
+import useTheme from "@/hooks/useTheme";
+
+type Props = {
+    label?: string;
+    icon?: keyof typeof Ionicons.glyphMap;
+    value: string;
+    onChangeText: (text: string) => void;
+    placeholder: string;
+    secureTextEntry?: boolean;
+    style?: object;
+};
+
+export default function InputField({
+                                       label,
+                                       icon,
+                                       value,
+                                       onChangeText,
+                                       placeholder,
+                                       secureTextEntry,
+                                       style
+                                   }: Readonly<Props>) {
+    const colors = useTheme();
+
+    return (
+        <View style={[styles.container, style]}>
+            {/* Label nur anzeigen, wenn vorhanden */}
+            {label ? (
+                <Text style={[styles.label, {color: colors.title}]}>{label}</Text>
+            ) : null}
+
+            <View style={[
+                styles.inputWrapper,
+                {backgroundColor: colors.input, paddingLeft: icon ? 10 : 14}
+            ]}>
+                {/* Icon nur anzeigen, wenn vorhanden */}
+                {icon && (
+                    <Ionicons name={icon} size={20} color={colors.subtitle} style={styles.icon}/>
+                )}
+
+                <TextInput
+                    placeholder={placeholder}
+                    placeholderTextColor="#999"
+                    style={[styles.input, {color: colors.title}]}
+                    value={value}
+                    onChangeText={onChangeText}
+                    secureTextEntry={secureTextEntry}
+                />
+            </View>
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+    container: {marginBottom: 16},
+    label: {fontSize: 14, marginBottom: 6},
+    inputWrapper: {
+        flexDirection: "row",
+        alignItems: "center",
+        paddingVertical: 10,
+        paddingRight: 14,
+        borderRadius: 14,
+    },
+    icon: {marginRight: 10},
+    input: {flex: 1, fontSize: 16},
+});
